Refactor code to be reused by future workloads
